[~] Fixed some bugs in the timing

+++ Source/Events/EventManager.cpp
@@ -118,7 +118,7 @@ int EventManager::GetSamplesFromAllEvents(int numSamples, Frame<float> *buffer)
         generated += samplesToGet;
     }
     auto end = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> seconds = end-start;
+    std::chrono::duration<float> seconds = end-start;
     GameObject::SetParam("DSPLoad", seconds.count() / (512.0f/44100.0f));
     GameObject::SetParam("HRTFLoad", GameObject::GetParam<float>("HRTFLoadTemp") / (512.0f / 44100.0f));
     GameObject::SetParam("ITDLoad", GameObject::GetParam<float>("ITDLoadTemp") / (512.0f / 44100.0f));
